+++ packages/vite-plugin-svelte/src/index.ts
@@ -1,5 +1,7 @@
 import fs from 'fs';
 import { HmrContext, ModuleNode, Plugin, ResolvedConfig, UserConfig } from 'vite';
+// eslint-disable-next-line node/no-missing-import
+import { isDepExcluded } from 'vitefu';
 import { handleHotUpdate } from './handle-hot-update';
 import { log, logCompilerWarnings } from './utils/log';
 import { CompileData, createCompileSvelte } from './utils/compile';
@@ -13,7 +15,6 @@ import {
 	patchResolvedViteConfig,
 	preResolveOptions
 } from './utils/options';
-import { VitePluginSvelteCache } from './utils/vite-plugin-svelte-cache';
 
 import { ensureWatchedFile, setupWatchers } from './utils/watch';
 import { resolveViaPackageJsonSvelte } from './utils/resolve';
@@ -21,6 +22,7 @@ import { PartialResolvedId } from 'rollup';
 import { toRollupError } from './utils/error';
 import { saveSvelteMetadata } from './utils/optimizer';
 import { svelteInspector } from './ui/inspector/plugin';
+import { VitePluginSvelteCache } from './utils/vite-plugin-svelte-cache';
 
 interface PluginAPI {
 	/**
@@ -68,7 +70,7 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 				// @ts-expect-error temporarily lend the options variable until fixed in configResolved
 				options = await preResolveOptions(inlineOptions, config, configEnv);
 				// extra vite config
-				const extraViteConfig = buildExtraViteConfig(options, config);
+				const extraViteConfig = await buildExtraViteConfig(options, config);
 				log.debug('additional vite config', extraViteConfig);
 				return extraViteConfig;
 			},
@@ -154,21 +156,32 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 					}
 					return resolvedSvelteSSR;
 				}
-				try {
-					const resolved = resolveViaPackageJsonSvelte(importee, importer, cache);
-					if (resolved) {
-						log.debug(
-							`resolveId resolved ${resolved} via package.json svelte field of ${importee}`
+				//@ts-expect-error scan
+				const scan = !!opts?.scan; // scanner phase of optimizeDeps
+				const isPrebundled =
+					options.prebundleSvelteLibraries &&
+					viteConfig.optimizeDeps?.disabled !== true &&
+					viteConfig.optimizeDeps?.disabled !== (options.isBuild ? 'build' : 'dev') &&
+					!isDepExcluded(importee, viteConfig.optimizeDeps?.exclude ?? []);
+				// for prebundled libraries we let vite resolve the prebundling result
+				// for ssr, during scanning and non-prebundled, we do it
+				if (ssr || scan || !isPrebundled) {
+					try {
+						const resolved = await resolveViaPackageJsonSvelte(importee, importer, cache);
+						if (resolved) {
+							log.debug(
+								`resolveId resolved ${resolved} via package.json svelte field of ${importee}`
+							);
+							return resolved;
+						}
+					} catch (e) {
+						log.debug.once(
+							`error trying to resolve ${importee} from ${importer} via package.json svelte field `,
+							e
 						);
-						return resolved;
+						// this error most likely happens due to non-svelte related importee/importers so swallow it here
+						// in case it really way a svelte library, users will notice anyway. (lib not working due to failed resolve)
 					}
-				} catch (e) {
-					log.debug.once(
-						`error trying to resolve ${importee} from ${importer} via package.json svelte field `,
-						e
-					);
-					// this error most likely happens due to non-svelte related importee/importers so swallow it here
-					// in case it really way a svelte library, users will notice anyway. (lib not working due to failed resolve)
 				}
 			},
 
@@ -215,6 +228,9 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 						throw toRollupError(e, options);
 					}
 				}
+			},
+			async buildEnd() {
+				await options.stats?.finishAll();
 			}
 		}
 	];