+++ packages/vite-plugin-svelte/src/utils/esbuild.ts
@@ -1,13 +1,16 @@
-import { promises as fs } from 'fs';
+import { readFileSync } from 'fs';
 import { compile, preprocess } from 'svelte/compiler';
 import { DepOptimizationOptions } from 'vite';
 import { Compiled } from './compile';
 import { log } from './log';
 import { CompileOptions, ResolvedOptions } from './options';
 import { toESBuildError } from './error';
+import { atLeastSvelte } from './svelte-version';
+import { StatCollection } from './vite-plugin-svelte-stats';
 
 type EsbuildOptions = NonNullable<DepOptimizationOptions['esbuildOptions']>;
 type EsbuildPlugin = NonNullable<EsbuildOptions['plugins']>[number];
+const isCssString = atLeastSvelte('3.53.0');
 
 export const facadeEsbuildSveltePluginName = 'vite-plugin-svelte:facade';
 
@@ -21,15 +24,23 @@ export function esbuildSveltePlugin(options: ResolvedOptions): EsbuildPlugin {
 
 			const svelteExtensions = (options.extensions ?? ['.svelte']).map((ext) => ext.slice(1));
 			const svelteFilter = new RegExp(`\\.(` + svelteExtensions.join('|') + `)(\\?.*)?$`);
-
+			let statsCollection: StatCollection | undefined;
+			build.onStart(() => {
+				statsCollection = options.stats?.startCollection('prebundle libraries', {
+					logResult: (c) => c.stats.length > 1
+				});
+			});
 			build.onLoad({ filter: svelteFilter }, async ({ path: filename }) => {
-				const code = await fs.readFile(filename, 'utf8');
+				const code = readFileSync(filename, 'utf8');
 				try {
-					const contents = await compileSvelte(options, { filename, code });
+					const contents = await compileSvelte(options, { filename, code }, statsCollection);
 					return { contents };
 				} catch (e) {
 					return { errors: [toESBuildError(e, options)] };
 				}
+			});
+			build.onEnd(() => {
+				statsCollection?.finish();
 			});
 		}
 	};
@@ -37,11 +48,16 @@ export function esbuildSveltePlugin(options: ResolvedOptions): EsbuildPlugin {
 
 async function compileSvelte(
 	options: ResolvedOptions,
-	{ filename, code }: { filename: string; code: string }
+	{ filename, code }: { filename: string; code: string },
+	statsCollection?: StatCollection
 ): Promise<string> {
+	let css = options.compilerOptions.css;
+	if (css !== 'none') {
+		css = isCssString ? 'injected' : true;
+	}
 	const compileOptions: CompileOptions = {
 		...options.compilerOptions,
-		css: true,
+		css,
 		filename,
 		format: 'esm',
 		generate: 'dom'
@@ -77,8 +93,10 @@ async function compileSvelte(
 				...dynamicCompileOptions
 		  }
 		: compileOptions;
-
+	const endStat = statsCollection?.start(filename);
 	const compiled = compile(finalCode, finalCompileOptions) as Compiled;
-
+	if (endStat) {
+		endStat();
+	}
 	return compiled.js.code + '//# sourceMappingURL=' + compiled.js.map.toUrl();
 }